pyproject.toml
[project]
name = "tg-bot-weather"
version = "0.1.0"
description = "Weather Telegram bot. Current conditions, 5-day forecast, favorites, inline-mode. OpenWeather API."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "jot1rsio" }]
keywords = ["telegram", "bot", "aiogram", "weather", "openweather", "template"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Framework :: AsyncIO",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
]

dependencies = [
    "aiogram>=3.13,<4.0",
    "pydantic>=2.9,<3.0",
    "pydantic-settings>=2.6,<3.0",
    "sqlalchemy[asyncio]>=2.0.36,<3.0",
    "aiosqlite>=0.20.0",
    "aiohttp>=3.10,<4.0",
    "fluent.runtime>=0.4.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "aioresponses>=0.7.6",
]

[project.urls]
Homepage = "https://github.com/arsenshahverdyanlavtxa228/tg-bot-weather"
Repository = "https://github.com/arsenshahverdyanlavtxa228/tg-bot-weather"
Issues = "https://github.com/arsenshahverdyanlavtxa228/tg-bot-weather/issues"

[project.scripts]
weather-bot = "bot.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bot"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W", "F", "I", "B", "C4", "UP", "SIM", "RUF", "N", "ASYNC", "PTH", "TID",
]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "B017", "RUF001", "RUF002", "RUF003"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_return_any = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["fluent.*", "aioresponses.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q --strict-markers --cov=bot --cov-report=term-missing"
